Namespace ArelSpecHelper to Spec::Support::ArelHelper
